Remove unnecessary casts
Clippy complains about unnecessary casts, since casting to the same
type is redundant.

+++ src/server.rs
@@ -563,7 +563,7 @@ impl<F: FileSystem + Sync> Server<F> {
         let Rename2In { newdir, flags, .. } = r.read_obj().map_err(Error::DecodeMessage)?;
 
         let flags =
-            flags & (libc::RENAME_EXCHANGE | libc::RENAME_NOREPLACE | libc::RENAME_WHITEOUT) as u32;
+            flags & (libc::RENAME_EXCHANGE | libc::RENAME_NOREPLACE | libc::RENAME_WHITEOUT);
 
         self.do_rename(in_header, size_of::<Rename2In>(), newdir, flags, r, w)
     }
